Relocate utils and ast_utils
