+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/azle_ast/mod.rs
@@ -7,3 +7,7 @@ pub enum CanisterMethodType {
     Query,
     Update,
 }
+
+pub enum SystemStructureType {
+    Canister,
+}
