+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/ts_ast/ident.rs
