+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/ts_ast/mod.rs
@@ -1,4 +1,5 @@
 pub mod fn_decl;
+pub mod ident;
 mod module;
 pub mod program;
 mod ts_array_type;
