+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/ts_ast/module.rs
@@ -1,4 +1,8 @@
-use swc_ecma_ast::{ExportDecl, Module, ModuleDecl};
+use swc_ecma_ast::{ExportDecl, Module, ModuleDecl, TsTypeAliasDecl};
+
+use crate::generators::canister_methods::{
+    get_ast_type_alias_decls_by_type_ref_name, get_type_alias_decls,
+};
 
 pub fn get_export_decls(module: &Module) -> Vec<ExportDecl> {
     let module_decls: Vec<ModuleDecl> = module
@@ -15,4 +19,10 @@ pub fn get_export_decls(module: &Module) -> Vec<ExportDecl> {
         .collect();
 
     export_decls
+}
+
+pub fn get_ast_func_type_alias_decls(module: &Module) -> Vec<TsTypeAliasDecl> {
+    // TODO should we have this take a Vec<TsTypeAliasDecl> so its more like the other ones like it in mod.rs and we should all put them in the same place, probably here.
+    let decls = get_type_alias_decls(module);
+    get_ast_type_alias_decls_by_type_ref_name(&decls, "Func")
 }
