+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/ts_ast/program.rs
@@ -1,7 +1,32 @@
-use swc_ecma_ast::{FnDecl, Program};
+use swc_ecma_ast::{FnDecl, Program, TsTypeAliasDecl};
 
-use super::{fn_decl::is_canister_method_type_fn_decl, module::get_export_decls};
-use crate::azle_ast::CanisterMethodType;
+use super::{
+    fn_decl::is_canister_method_type_fn_decl, module, module::get_export_decls,
+    ts_type_alias_decl::is_type_alias_decl_system_structure_type,
+};
+use crate::{
+    azle_ast::{CanisterMethodType, SystemStructureType},
+    generators::canister_methods::get_ast_type_alias_decls_from_programs,
+};
+
+// TODO: This may grab unintended Func declarations. Instead traverse starting
+// from the exported canister methods.
+pub fn get_ast_func_type_alias_decls_from_programs(
+    programs: &Vec<Program>,
+) -> Vec<TsTypeAliasDecl> {
+    programs.iter().fold(vec![], |acc, program| {
+        let ast_func_type_alias_decls = get_ast_func_type_alias_decls_from_program(program);
+
+        vec![acc, ast_func_type_alias_decls].concat()
+    })
+}
+
+pub fn get_ast_func_type_alias_decls_from_program(program: &Program) -> Vec<TsTypeAliasDecl> {
+    match program {
+        Program::Module(module) => module::get_ast_func_type_alias_decls(module),
+        Program::Script(_) => vec![],
+    }
+}
 
 pub fn get_canister_method_type_fn_decls(
     programs: &Vec<Program>,
@@ -40,4 +65,18 @@ fn get_ast_fn_decls_from_program(program: &Program) -> Vec<FnDecl> {
             vec![]
         }
     }
+}
+
+pub fn get_type_alias_decls_for_system_structure_type(
+    programs: &Vec<Program>,
+    system_structure_type: &SystemStructureType,
+) -> Vec<TsTypeAliasDecl> {
+    let type_alias_decls = get_ast_type_alias_decls_from_programs(programs);
+
+    type_alias_decls
+        .into_iter()
+        .filter(|type_alias_decl| {
+            is_type_alias_decl_system_structure_type(type_alias_decl, system_structure_type)
+        })
+        .collect()
 }
